requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-request-logging==0.7.5
drf-yasg==1.21.7